code-feat-production-grade-upgrade/requirements.txt
streamlit
pandas
ccxt
ccxt-pro
# The 'ccxt-pro' package is a commercial product and is not available on PyPI.
# It must be installed separately from the private GitHub repository after purchase.
# The application includes a mock version for development purposes if ccxt-pro is not found.
# pip install git+https://github.com/kroitor/ccxt.pro.git#subdirectory=python
web3
httpx
aiosqlite
plotly
python-dotenv
faker
numpy
nest_asyncio
PyYAML
pytest
pytest-asyncio
pandas-ta
requests
pycoingecko
cachetools
